% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_breaks.R
\name{api_breaks}
\alias{api_breaks}
\title{Generate breaks for a Curbcut page}
\usage{
api_breaks(var_left, var_right = " ", scale, time, region = NULL)
}
\arguments{
\item{var_left}{<\code{character}>}

\item{var_right}{<\code{character}>}

\item{scale}{<\code{character}>}

\item{time}{<\code{character}>}

\item{region}{<\code{character}>}
}
\value{
A list containing three components.
}
\description{
Generate breaks for a Curbcut page
}
