% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_context.R
\name{context}
\alias{context}
\title{This function generates contextual information such as legend, graph, and text
for the application based on provided parameters. It fetches variables from
the database, processes them, and then generates the necessary components
for the application's context.}
\usage{
context(
  var_left,
  var_right = " ",
  scale,
  region = NULL,
  time,
  select_id,
  lang = NULL,
  top_scale,
  schemas = list(var_left = list(time = time), var_right = list(time = time))
)
}
\arguments{
\item{var_left}{<\code{character}>}

\item{var_right}{<\code{character}>}

\item{scale}{<\code{character}>}

\item{region}{<\code{character}>}

\item{time}{<\code{character}>}

\item{select_id}{<\code{character}>}

\item{lang}{<\verb{character vector}>}

\item{top_scale}{<\verb{character vector}>}

\item{schemas}{<\code{list}>}
}
\value{
A list containing three components.
}
\description{
This function generates contextual information such as legend, graph, and text
for the application based on provided parameters. It fetches variables from
the database, processes them, and then generates the necessary components
for the application's context.
}
