% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta_fun}
\alias{data_get_delta_fun}
\alias{data_get_delta_fun.scalar}
\alias{data_get_delta_fun.ordinal}
\title{Inner function to get data based on the type of \code{vars}}
\usage{
data_get_delta_fun(
  vars,
  scale,
  region,
  variables,
  schema,
  breaks,
  reduce = TRUE,
  time,
  ...
)

\method{data_get_delta_fun}{scalar}(
  vars,
  scale,
  region,
  variables,
  schema,
  breaks,
  time,
  reduce = TRUE,
  ...
)

\method{data_get_delta_fun}{ordinal}(
  vars,
  scale,
  region,
  variables,
  schema,
  breaks,
  reduce = TRUE,
  time,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{scale}{<\code{character}> The scale at which the user is on.}

\item{region}{<\verb{character vector}> String of the region under study}

\item{variables}{<\code{data.frame}> Dataframe of the variables dictionary, containing
both var_left and var_right, and any potential parent variable aswell.}

\item{breaks}{<\code{vector}>}

\item{reduce}{<\code{logical}> Should the dataframe be reduced to a single table
(if there are multiple scales)}

\item{time}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function. It contains the time for both var_left and var_right variables.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This function dispatches the data retrieval based on the class
of the \code{vars} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data_get_delta_fun(scalar)}: The method for scalar variables.

\item \code{data_get_delta_fun(ordinal)}: The method for ordinal variables.

}}
\seealso{
\code{\link{data_get.delta}}
}
