% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_db_conn.R
\name{db_connection}
\alias{db_connection}
\title{Establish a Database Connection Pool}
\usage{
db_connection()
}
\value{
<\code{DBI::dbPool}> A database connection pool object.
}
\description{
This function creates a database connection pool using the \code{pool} package
and connects to the prod PostgreSQL database using environment variables for
configuration.
}
