% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{db_get_data_from_sql}
\alias{db_get_data_from_sql}
\title{Retrieve data from a QS file based on variable and scale}
\usage{
db_get_data_from_sql(vars_vector, scales, region, schema)
}
\arguments{
\item{vars_vector}{<\code{character}> A string specifying the name of the table to
retrieve data from. Single variable = single table. e.g. \code{housing_tenant}}

\item{region}{<\code{character}> String of the region under study.}

\item{schema}{<\code{public}>}

\item{scale}{<\code{character}> A string specifying the scale at which to retrieve
data, corresponding to a path on disk, e.g. \code{DA} or \code{CSD}.}
}
\value{
A data.frame object with the selected data from the specified table.
}
\description{
This function takes in a variable code and the scale to retrieve data from a
QS file. The function constructs the file path based on the input, then reads
the data using the qs package's \code{\link[qs]{qread}} function.
}
