% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_data_present_in_scale}
\alias{is_data_present_in_scale}
\title{Check if data is present in specified scale}
\usage{
is_data_present_in_scale(var, scale, variables)
}
\arguments{
\item{var}{<\code{character}> The variable name to check for in the scale's file list.}

\item{scale}{<\code{character}> The name of the scale for which the data presence
is to be checked. It is expected to correspond to a variable containing a list
of file names.}

\item{variables}{<\code{table}>}
}
\value{
<\code{logical}> Returns \code{TRUE} if the variable and scale is found together
in the variables table
}
\description{
This function checks if a particular variable is present within a specified scale.
}
