% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_context.R
\name{run_api}
\alias{run_api}
\title{Run Plumber API}
\usage{
run_api()
}
\value{
This function does not return a value. It starts the Plumber API server.
}
\description{
This function initializes and runs the Plumber API for the \code{cc.Rapi} package.
It also sets up a hook to close the database pool when the API is stopped.
}
